requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
Flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3